requirements.txt
phonenumbers==8.9.14
pyserial==3.4
